+++ packages/spacecat-shared-data-access/test/unit/models/audit/audit.model.test.js
@@ -177,11 +177,14 @@ describe('AuditModel', () => {
       ACCESSIBILITY: 'accessibility',
       SECURITY_CSP: 'security-csp',
       PAID: 'paid',
+      HREFLANG: 'hreflang',
+      PAID_TRAFFIC_ANALYSIS_WEEKLY: 'paid-traffic-analysis-weekly',
+      PAID_TRAFFIC_ANALYSIS_MONTHLY: 'paid-traffic-analysis-monthly',
     };
 
     it('should have all audit types present in AUDIT_TYPES', () => {
       expect(auditTypes).to.eql(expectedAuditTypes);
-      expect(Object.keys(auditTypes)).to.have.lengthOf(26);
+      expect(Object.keys(auditTypes)).to.have.lengthOf(29);
     });
 
     it('should not have unexpected audit types in AUDIT_TYPES', () => {
@@ -239,6 +242,7 @@ describe('AuditModel', () => {
       const stepResult = {
         type: 'someType',
         siteId: 'someSiteId',
+        endDate: '2025-08-12T15:46:00.000Z',
         urlConfigs: [{ url: 'someUrl', geo: 'someGeo' }],
       };
       const auditContext = { some: 'context' };
@@ -249,6 +253,8 @@ describe('AuditModel', () => {
         type: 'someType',
         siteId: 'someSiteId',
         pageUrl: undefined,
+        startDate: undefined,
+        endDate: '2025-08-12T15:46:00.000Z',
         urlConfigs: [{ url: 'someUrl', geo: 'someGeo' }],
         allowCache: true,
         auditContext: { some: 'context' },