+++ packages/spacecat-shared-data-access/test/unit/models/site/site.model.test.js
@@ -62,7 +62,7 @@ describe('SiteModel', () => {
       const result = computeExternalIds(attrs, Site.AUTHORING_TYPES);
 
       expect(result).to.deep.equal({
-        externalOwnerId: 'main#adobe',
+        externalOwnerId: 'adobe',
         externalSiteId: 'example-site',
       });
     });
@@ -96,7 +96,7 @@ describe('SiteModel', () => {
       const result = computeExternalIds(attrs, Site.AUTHORING_TYPES);
 
       expect(result).to.deep.equal({
-        externalOwnerId: 'main#adobe',
+        externalOwnerId: 'adobe',
         externalSiteId: undefined,
       });
     });